+++ lib/bmg/relation.rb
@@ -21,6 +21,12 @@ module Bmg
       Bmg::Type::ANY
     end
 
+    def with_type(type)
+      dup.tap{|r|
+        r.type = type
+      }
+    end
+
     def with_typecheck
       dup.tap{|r|
         r.type = r.type.with_typecheck